services/ai-ml-service/app/services/ml_service.py
"""
ML Service for handling machine learning operations

Demonstrates AI/ML capabilities for XYZ Dev Foundation:
- TensorFlow model loading and inference
- Scikit-learn algorithms
- Real-time predictions
- Model management and versioning
"""

import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
import tensorflow as tf
import logging
import asyncio
from typing import Dict, List, Any, Optional
import pickle
import json
from datetime import datetime

logger = logging.getLogger(__name__)

class MLService:
    def __init__(self):
        self.models = {}
        self.scalers = {}
        self.models_loaded = False
        
    async def load_models(self):
        """Load pre-trained ML models"""
        try:
            logger.info("Loading ML models...")
            
            # Load recommendation model
            self.models['recommendation'] = self._create_recommendation_model()
            
            # Load sentiment analysis model
            self.models['sentiment'] = self._create_sentiment_model()
            
            # Load classification model
            self.models['classification'] = self._create_classification_model()
            
            # Load regression model
            self.models['regression'] = self._create_regression_model()
            
            self.models_loaded = True
            logger.info("All ML models loaded successfully")
            
        except Exception as e:
            logger.error(f"Failed to load ML models: {str(e)}")
            raise
    
    async def predict(self, model_name: str, features: List[float], context: Optional[Dict] = None) -> Dict[str, Any]:
        """Make predictions using specified model"""
        if not self.models_loaded:
            raise ValueError("Models not loaded")
            
        if model_name not in self.models:
            raise ValueError(f"Model {model_name} not found")
        
        try:
            model = self.models[model_name]
            features_array = np.array(features).reshape(1, -1)
            
            if model_name in ['classification', 'regression']:
                # Scikit-learn models
                prediction = model.predict(features_array)[0]
                if hasattr(model, 'predict_proba'):
                    confidence = float(np.max(model.predict_proba(features_array)))
                else:
                    confidence = 0.85  # Default confidence for regression
            else:
                # Custom models
                prediction = await self._custom_predict(model_name, features_array, context)
                confidence = 0.90
            
            return {
                "prediction": prediction,
                "confidence": confidence,
                "model_version": "1.0.0",
                "features_used": len(features)
            }
            
        except Exception as e:
            logger.error(f"Prediction failed for model {model_name}: {str(e)}")
            raise
    
    async def generate_recommendations(self, user_id: str, item_type: str, 
                                    num_recommendations: int, filters: Optional[Dict] = None) -> List[Dict]:
        """Generate personalized recommendations"""
        try:
            # Simulate recommendation generation
            recommendations = []
            for i in range(num_recommendations):
                recommendations.append({
                    "item_id": f"{item_type}_{i+1}",
                    "title": f"Recommended {item_type.title()} {i+1}",
                    "score": round(0.95 - (i * 0.05), 2),
                    "category": item_type,
                    "reason": "Based on your preferences and similar users"
                })
            
            return recommendations
            
        except Exception as e:
            logger.error(f"Recommendation generation failed: {str(e)}")
            raise
    
    async def analyze_sentiment(self, text: str, language: str = "en") -> Dict[str, Any]:
        """Analyze sentiment of text"""
        try:
            # Simple sentiment analysis simulation
            positive_words = ["good", "great", "excellent", "amazing", "wonderful", "fantastic"]
            negative_words = ["bad", "terrible", "awful", "horrible", "disappointing"]
            
            text_lower = text.lower()
            positive_count = sum(1 for word in positive_words if word in text_lower)
            negative_count = sum(1 for word in negative_words if word in text_lower)
            
            if positive_count > negative_count:
                sentiment = "positive"
                confidence = min(0.9, 0.6 + (positive_count * 0.1))
            elif negative_count > positive_count:
                sentiment = "negative"
                confidence = min(0.9, 0.6 + (negative_count * 0.1))
            else:
                sentiment = "neutral"
                confidence = 0.7
            
            return {
                "sentiment": sentiment,
                "confidence": confidence,
                "emotions": {
                    "joy": 0.8 if sentiment == "positive" else 0.2,
                    "sadness": 0.8 if sentiment == "negative" else 0.2,
                    "anger": 0.1,
                    "fear": 0.1
                }
            }
            
        except Exception as e:
            logger.error(f"Sentiment analysis failed: {str(e)}")
            raise
    
    async def get_available_models(self) -> List[Dict[str, Any]]:
        """Get list of available models"""
        models_info = []
        for model_name in self.models.keys():
            models_info.append({
                "name": model_name,
                "type": self._get_model_type(model_name),
                "version": "1.0.0",
                "status": "active",
                "last_updated": datetime.now().isoformat()
            })
        return models_info
    
    async def get_model_info(self, model_name: str) -> Optional[Dict[str, Any]]:
        """Get detailed information about a model"""
        if model_name not in self.models:
            return None
            
        return {
            "name": model_name,
            "type": self._get_model_type(model_name),
            "version": "1.0.0",
            "description": f"ML model for {model_name} tasks",
            "input_features": self._get_model_features(model_name),
            "output_format": "prediction with confidence score",
            "performance_metrics": {
                "accuracy": 0.92,
                "precision": 0.89,
                "recall": 0.91
            },
            "created_at": "2024-01-01T00:00:00Z",
            "last_updated": datetime.now().isoformat()
        }
    
    async def log_recommendation_event(self, user_id: str, recommendations: List[Dict]):
        """Log recommendation event for analytics"""
        try:
            event = {
                "user_id": user_id,
                "event_type": "recommendations_generated",
                "recommendations_count": len(recommendations),
                "timestamp": datetime.now().isoformat()
            }
            logger.info(f"Logged recommendation event: {event}")
        except Exception as e:
            logger.error(f"Failed to log recommendation event: {str(e)}")
    
    def _create_recommendation_model(self):
        """Create a simple recommendation model"""
        # Simulate a trained recommendation model
        return {"type": "collaborative_filtering", "version": "1.0.0"}
    
    def _create_sentiment_model(self):
        """Create a sentiment analysis model"""
        return {"type": "nlp_sentiment", "version": "1.0.0"}
    
    def _create_classification_model(self):
        """Create a classification model"""
        # Create a simple RandomForest classifier
        X = np.random.rand(1000, 10)
        y = np.random.randint(0, 3, 1000)
        model = RandomForestClassifier(n_estimators=100, random_state=42)
        model.fit(X, y)
        return model
    
    def _create_regression_model(self):
        """Create a regression model"""
        # Create a simple RandomForest regressor
        X = np.random.rand(1000, 5)
        y = np.random.rand(1000)
        model = RandomForestRegressor(n_estimators=100, random_state=42)
        model.fit(X, y)
        return model
    
    async def _custom_predict(self, model_name: str, features: np.ndarray, context: Optional[Dict]) -> Any:
        """Handle custom model predictions"""
        if model_name == "recommendation":
            return {"recommended_items": [1, 2, 3, 4, 5]}
        elif model_name == "sentiment":
            return "positive"
        else:
            return np.random.rand()
    
    def _get_model_type(self, model_name: str) -> str:
        """Get model type"""
        type_mapping = {
            "classification": "classification",
            "regression": "regression", 
            "recommendation": "recommendation",
            "sentiment": "nlp"
        }
        return type_mapping.get(model_name, "unknown")
    
    def _get_model_features(self, model_name: str) -> List[str]:
        """Get model input features"""
        feature_mapping = {
            "classification": ["feature_1", "feature_2", "feature_3"],
            "regression": ["value_1", "value_2", "value_3"],
            "recommendation": ["user_id", "item_features"],
            "sentiment": ["text"]
        }
        return feature_mapping.get(model_name, [])